Fixes documentation. Renames isMounted.ts to useIsMounted.ts.
