+++ .prettierignore
@@ -1,2 +1,1 @@
 /build
-/boilerplate
