+++ boilerplate/app/screens/WelcomeScreen.tsx
@@ -1,7 +1,10 @@
 import { observer } from "mobx-react-lite"
 import React, { FC } from "react"
 import { Image, ImageStyle, TextStyle, View, ViewStyle } from "react-native"
-import { Button, Text } from "../components"
+import {
+  Button, // @demo remove-current-line
+  Text,
+} from "../components"
 import { isRTL } from "../i18n"
 import { useStores } from "../models" // @demo remove-current-line
 import { AppStackScreenProps } from "../navigators" // @demo remove-current-line
