+++ boilerplate/app/utils/useIsMounted.ts
